api/requirements.txt
fastapi
uvicorn
numpy<2.0
scikit-surprise
pandas
pydantic
